C#: Handle multiple-field Boolean CFG splitting
The internal pre-SSA library was extended on 413ef7ec0999841375ab6af416aa06c69157c4be
to include fields/properties that are local-scope-like. The CFG splitting logic
uses ranking of SSA definitions to define an (arbitrary) order of splits, but for
fields/properties the implicit entry definition all have the same line and column.
In effect, such SSA definitions incorrectly get the same rank. Adding the name
of the field/property to the lexicographic ordering resolves the issue.