C#: Revert recent change to `AccessorCall`
The recent change to `AccessorCall` on 34e6173a221cce81174582a9d182db9f7403a3e4 resulted
in some bad join-orders, so I have (partly) reverted them. This means that the issues
orignally addressed by that change are now reintroduced, and I plan to instead apply a
fix to the CFG, which--unlike the original fix--should be able to handle multi-property-tuple
assignments.